+++ src/lib/libft/src/ft_memcmp.c
@@ -13,7 +13,7 @@
 
 int	ft_memcmp(const void *ptr_find, const void *ptr_str, size_t length)
 {
-	int				index;
+	size_t			index;
 	unsigned char	*aux1;
 	unsigned char	*aux2;
 
@@ -22,7 +22,7 @@ int	ft_memcmp(const void *ptr_find, const void *ptr_str, size_t length)
 	index = 0;
 	aux1 = (unsigned char *) ptr_find;
 	aux2 = (unsigned char *) ptr_str;
-	while ((aux2[index] != 0 || aux1[index] != 0) && index < (int) length)
+	while ((aux2[index] != 0 || aux1[index] != 0) && index < length)
 	{
 		if (aux2[index] != aux1[index])
 			return (aux1[index] - aux2[index]);
