+++ src/lib/libft/src/ft_strnstr.c
@@ -11,12 +11,12 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-static	int	ft_fndstr(const char *str1, const char *substr, size_t l, int i)
+static	int	ft_fndstr(const char *str1, const char *substr, size_t l, size_t i)
 {
-	int	j;
+	size_t	j;
 
 	j = 0;
-	while (i < (int) l)
+	while (i < l)
 	{
 		if (str1[i] != substr[j])
 			break ;
@@ -30,7 +30,7 @@ static	int	ft_fndstr(const char *str1, const char *substr, size_t l, int i)
 
 char	*ft_strnstr(const char *string, const char *to_find, size_t length)
 {
-	int				index;
+	size_t			index;
 	unsigned char	*ptr_found;
 
 	if (!string || !to_find)
@@ -39,7 +39,7 @@ char	*ft_strnstr(const char *string, const char *to_find, size_t length)
 	if (to_find[0] == 0)
 		return ((char *) ptr_found);
 	index = 0;
-	while (index < (int) length && string[index] != 0)
+	while (index < length && string[index] != 0)
 	{
 		if (string[index] == to_find[0])
 			if (ft_fndstr(string, to_find, length, index))
